backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
dotenv
openai
PyMuPDF
python-pptx
SQLAlchemy
psycopg2-binary
python_pptx
